Adds client logos and section
Had to fiddle around with the templates to add the client section.

Still waiting for a better Piton logo from Max.

Still need to write short case studies for Piton and Rocket Science

Decreased font sizes a bit so logos show up above the fold on homepage.

Still trying to capture "what we do" (or want to do) in the homepage copy.